+++ Cargo.toml
@@ -26,6 +26,7 @@ ndarray = "0.16.1"
 hashbrown = "0.15.2"
 getset = "0.1.3"
 parking_lot = { version = "0.12.3", optional = true }
+molly = { version = "0.3.2", optional = true }
 
 [dev-dependencies]
 file_diff = "1.0.0"
@@ -40,6 +41,7 @@ cc = { version = "1.2.1", features = ["parallel"] }
 [features]
 serde = ["ndarray/serde", "hashbrown/serde"]
 parallel = ["parking_lot"]
+molly = ["dep:molly"]
 
 [[bench]]
 name = "main"
