+++ tmp_xtc_iter_range_alt_print_progress.txt
@@ -1,0 +1,1 @@
+[ RUNNING ]   Step        20000 | Time          400 ps[ RUNNING ]   Step        35000 | Time          700 ps[COMPLETED]   Step        40000 | Time          800 ps
