+++ tmp_xtc_iter_step_range_print_progress.txt
@@ -1,0 +1,1 @@
+[ JUMPING ]   Jumping to the start of the iteration...[ RUNNING ]   Step        20000 | Time          400 ps[ RUNNING ]   Step        35000 | Time          700 ps[COMPLETED]   Step        35000 | Time          700 ps
